Initial commit for siasync version 1.0
siasync now works with 1.4.0+ and has many new features that can be
found with `siasync -h`.  It also does some checking with `siad` that
there are contracts and allowance available to upload files.

There are a number of bugs still around.  The deletion of files is
confusing when using the include or exclude flags.  More improvement is
needed in the Readme documentation as well.